requirements.txt
django
djangorestframework
django-cors-headers
django-filter
drf-yasg
django-environ
djangorestframework-simplejwt
whitenoise
dotenev
pandas